Исправлены некоторые ошибки
