+++ src/components/ui/NoContent/styles.js
@@ -3,4 +3,5 @@ import styled from "styled-components";
 export const NoContent = styled.p`
   margin: 0 auto;
   font-size: 20px;
+  text-align: center;
 `;
