+++ src/pages/Auth/Inputs/index.jsx
@@ -1,15 +1,8 @@
 import { FormElement } from "../../../components/ui/FormElement";
 import { Input } from "../../../components/ui/Input";
 import { Label } from "../../../components/ui/Label";
-import * as SC from "./styles";
 
-export const Inputs = ({
-  onChange,
-  isAdmin,
-  handleChange,
-  isGuest,
-  setIsGuest,
-}) => (
+export const Inputs = ({ onChange }) => (
   <>
     <FormElement>
       <Input
@@ -28,24 +21,5 @@ export const Inputs = ({
       />
       <Label htmlFor="password" text="Пароль" />
     </FormElement>
-
-    <SC.Label>
-      <input
-        type="checkbox"
-        id="checkbox"
-        checked={isAdmin}
-        onChange={() => handleChange()}
-      />
-      <label htmlFor="checkbox">Войти как администратор</label>
-    </SC.Label>
-    <SC.Label>
-      <input
-        type="checkbox"
-        id="guest"
-        checked={isGuest}
-        onChange={() => setIsGuest(!isGuest)}
-      />
-      <label htmlFor="guest">Войти как гость</label>
-    </SC.Label>
   </>
 );
