+++ src/pages/MyPage/Friends/Friend/styles.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const Friend = styled.div`
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+`;
