+++ src/pages/MyPage/Posts/Post/styles.js
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const Post = styled.div`
+  max-width: 250px;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  gap: 15px;
+  text-align: center;
+  padding: 10px 20px;
+  border: 2px solid #6261ba;
+  border-radius: 15px;
+  word-wrap: break-word;
+`;
