+++ src/pages/MyPage/Posts/index.jsx
@@ -1,27 +1,38 @@
-import { useState } from "react";
-import { Button } from "../../../components/ui/Button";
-import { Edit } from "./Edit";
+import { useSelector } from "react-redux";
+import { NoContent } from "../../../components/ui/NoContent";
+import { Post } from "./Post";
 import * as SC from "./styles";
 
-export const Posts = ({ myPost }) => {
-  const { id, post } = myPost;
-  const [postData, setPostData] = useState({});
-  const [edit, setEdit] = useState(false);
-
-  const editPost = (values) => {
-    setPostData(values);
-    setEdit(true);
-    return;
-  };
+export const Posts = () => {
+  const { user } = useSelector((state) => state.currentUser.currentUser);
+  const { posts } = useSelector((state) => state.posts.privatePosts);
+  const { publicPosts } = useSelector((state) => state.posts);
+  const filteredPosts = publicPosts?.filter((item) => {
+    if (item.user.email === user.email) {
+      return item;
+    }
+  });
   return (
-    <>
-      <SC.Post key={id}>
-        <p>{post}</p>
-        <Button onClick={() => editPost({ id, post })}>
-          Редактировать пост
-        </Button>
-      </SC.Post>
-      {edit && <Edit values={postData} setEdit={setEdit} />}
-    </>
+    <SC.SectionPosts>
+      <h3>Общедоступные посты</h3>
+      <SC.MyPosts>
+        {filteredPosts ? (
+          filteredPosts.map((filteredPost) => (
+            <Post key={filteredPost.id} myPost={filteredPost} />
+          ))
+        ) : (
+          <NoContent>Нет таких постов</NoContent>
+        )}
+      </SC.MyPosts>
+
+      <h3>Посты только для друзей</h3>
+      <SC.MyPosts>
+        {posts ? (
+          posts.map((post) => <Post key={post.id} myPost={post} />)
+        ) : (
+          <NoContent>Нет таких постов</NoContent>
+        )}
+      </SC.MyPosts>
+    </SC.SectionPosts>
   );
 };
