--- src/pages/MyPage/styles.js
@@ -1,35 +1,0 @@
-import styled from "styled-components";
-
-export const Info = styled.div`
-  display: flex;
-  gap: 5px;
-  flex-direction: column;
-  align-items: center;
-`;
-
-export const Name = styled.p`
-  font-size: 25px;
-`;
-
-export const Email = styled.p`
-  font-size: 15px;
-`;
-
-export const Friends = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  gap: 20px;
-`;
-
-export const MyPosts = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  gap: 30px;
-  margin: 20px 0;
-`;
-
-export const SectionPosts = styled.div`
-  display: flex;
-  flex-direction: column;
-  margin: 20px 0;
-`;
